+++ ui/src/components/mobileservices/BindingPanel.js
@@ -10,7 +10,7 @@ import '../configuration/ServiceSDKInfo.css';
 import './ServiceRow.css';
 import Form from "react-jsonschema-form";
 import debounce from 'lodash/debounce';
-
+import {OpenShiftObjectTemplate} from './bindingPanelUtils'
 
 class BindingPanel extends Component {
     
@@ -147,53 +147,6 @@ onBackButtonClick() {
 
 }
 
-//This renders the json fields using the rules from the form definition.
-function OpenShiftObjectTemplate({ TitleField, uiSchema, properties, title, description }) {
-  const form = uiSchema.form
-
-  return (
-    <div>
-    <TitleField title={title} />
-    <div className="row">
-      {form.map(key => (
-        getOpenShiftField(key, properties)
-      ))}
-    </div>
-    {description}
-  </div>
-  );
-}
-
-function getOpenShiftField(field, properties) {
-  if (!field.items) {
-    const property = properties.find(prop=>prop.name === field)
-    return <div
-              key={property.content.key}>
-              {property.content}
-            </div>
-  } else {
-    return getFieldSet(field, properties);
-  }
-}
-
-function getFieldSet(field, properties) {
-  const title = field.title;
-  const items = field.items;
-
-  const fieldSetItems = items.map(item => {
-    if (typeof item === "string") {
-      return properties.find(prop=>prop.name === item).content
-    } else {
-      return properties.find(prop=>prop.name === item.key).content
-    }
-  })
-
-  return <fieldset>
-    <h2>{title}</h2>
-  {
-    fieldSetItems
-  }</fieldset>;
-}
 
 
 const mapDispatchToProps = {
